Rename component to original name, remove confusing container
