+++ app/shared/resource-card/ResourceCard.js
