+++ app/shared/resource-card/ResourceCard.spec.js
@@ -11,12 +11,12 @@ import Unit from 'utils/fixtures/Unit';
 import { getResourcePageUrlComponents } from 'utils/resourceUtils';
 import { shallowWithIntl } from 'utils/testUtils';
 import ResourceAvailability from './label/ResourceAvailability';
-import { UnconnectedResourceCard } from './ResourceCardContainer';
+import { UnconnectedResourceCard } from './ResourceCard';
 import UnpublishedLabel from 'shared/label/Unpublished';
 import ResourceCardInfoCell from './info';
 import iconHeartWhite from 'assets/icons/heart-white.svg';
 
-describe('shared/resource-card/ResourceCardContainer', () => {
+describe('shared/resource-card/ResourceCard', () => {
   function getResource(extra) {
     return Immutable(
       Resource.build({
